Extract ConsistencyCheckJobItemProgressContext
