+++ kernel/data-pipeline/api/src/main/java/org/apache/shardingsphere/data/pipeline/api/job/progress/ConsistencyCheckJobItemProgress.java
@@ -28,6 +28,7 @@ import java.util.Map;
 /**
  * Data consistency check job item progress.
  */
+// TODO move package
 @Getter
 @RequiredArgsConstructor
 @ToString
