+++ kernel/data-pipeline/core/src/main/java/org/apache/shardingsphere/data/pipeline/core/job/progress/PipelineJobProgressDetector.java
@@ -47,6 +47,7 @@ public final class PipelineJobProgressDetector {
         return inventoryTasks.stream().allMatch(each -> each.getTaskProgress().getPosition() instanceof FinishedPosition);
     }
     
+    // TODO clean unused methods
     /**
      * Whether job is completed (successful or failed).
      *
