+++ kernel/data-pipeline/scenario/src/main/java/org/apache/shardingsphere/data/pipeline/scenario/consistencycheck/ConsistencyCheckTasksRunner.java
@@ -104,7 +104,7 @@ public final class ConsistencyCheckTasksRunner implements PipelineTasksRunner {
                 dataConsistencyCheckResult = jobAPI.dataConsistencyCheck(parentJobConfig, calculateAlgorithm, jobItemContext);
                 PipelineAPIFactory.getGovernanceRepositoryAPI().persistCheckJobResult(parentJobId, checkJobId, dataConsistencyCheckResult);
             } finally {
-                jobItemContext.setCheckEndTimeMillis(System.currentTimeMillis());
+                jobItemContext.getProgressContext().setCheckEndTimeMillis(System.currentTimeMillis());
             }
         }
